backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
numpy==1.26.3
opencv-python-headless==4.9.0.80
insightface==0.7.3
onnxruntime==1.16.3
google-cloud-firestore==2.14.0
google-auth==2.27.0
python-dotenv==1.0.1
